Remove PETS formatting
Dear author,

We regret to inform you that your submission #85 "Ethical identity,
ring VRFs, and..." to Issue 3 of PoPETs 2023 has not been accepted for
publication. The decision was “Desk Reject” and was made by the Chairs
without review by the PC.

The rationale behind the decision is explained in the short review that
you will find attached to this notification. If you would like to
resubmit to a future issue of PoPETs, please address the reason behind
the desk rejection.

Sincerely,
The PETS Program Co-Chairs
Michelle Mazurek and Micah Sherr

Review #85A
===========================================================================

Overall merit
-------------
1. Reject

Reviewer expertise
------------------
4. Expert

Reviewer confidence
-------------------
3. High

Relevance to PETS
-----------------
1. No/loose relation to PETs (justifies Reject)

Ethical considerations
----------------------
4. I don't see any ethical implications for this paper.

Comments for author
-------------------
The editors have determined that this paper falls outside the scope of PETS. A paper's applicability to building privacy enhancing technologies is crucial for our community. PETS welcomes proposals and improvements of privacy-preserving primitives as long as it is clear how the scheme or the improvement serves to create better privacy technologies (e.g., it includes a use case, evaluation on real data, integration on an application, etc.).