cargo test in the crate root fails without this
It worked in the individual crates before this, not sure why.
I bet this is due to ark_std::io being either std::io or the fork of it.

Awful lot of stupid shit in Rust is due to std::io not being core::io,
which afaik happens because of the missguided effort to add backtraces
everywhere.